# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pregnancies/mine:
    get:
      operationId: listMyPregnancies
      tags: [pregnancies]
      summary: List pregnancies the current person is a member of
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of pregnancy summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PregnancySummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pregnancies:
    post:
      operationId: createPregnancy
      tags: [pregnancies]
      summary: Create a new pregnancy and initial membership
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PregnancyInput"
      responses:
        "201":
          description: Created pregnancy and membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PregnancyWithMembership"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pregnancies/{id}:
    get:
      operationId: getPregnancy
      tags: [pregnancies]
      summary: Get a pregnancy with members
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pregnancy detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PregnancyDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a member of this pregnancy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Pregnancy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePregnancy
      tags: [pregnancies]
      summary: Update a pregnancy's name or due date
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PregnancyUpdate"
      responses:
        "200":
          description: Updated pregnancy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pregnancy"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a member of this pregnancy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Pregnancy:
      type: object
      required: [id, name, dueDatePrecision, status, joinCode, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date
        dueDatePrecision:
          type: string
          enum: [exact, approximate_month, unknown]
        status:
          type: string
        joinCode:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PregnancySummary:
      type: object
      required: [id, name, dueDatePrecision, status, joinCode, myRole, memberCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date
        dueDatePrecision:
          type: string
        status:
          type: string
        joinCode:
          type: string
        myRole:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    PregnancyInput:
      type: object
      required: [dueDatePrecision, role]
      properties:
        name:
          type: string
        dueDate:
          type: string
          format: date
        dueDatePrecision:
          type: string
          enum: [exact, approximate_month, unknown]
        role:
          type: string
          enum: [pregnant_person, supporter]
    PregnancyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        dueDate:
          type: string
          format: date
        dueDatePrecision:
          type: string
          enum: [exact, approximate_month, unknown]
    Membership:
      type: object
      required: [id, personId, pregnancyId, role, invitationStatus, createdAt]
      properties:
        id:
          type: integer
        personId:
          type: integer
        pregnancyId:
          type: integer
        role:
          type: string
        invitationStatus:
          type: string
        createdAt:
          type: string
          format: date-time
    MemberSummary:
      type: object
      required: [id, displayName, role]
      properties:
        id:
          type: integer
        displayName:
          type: string
        role:
          type: string
...
    PregnancyWithMembership:
      type: object
      required: [pregnancy, membership]
      properties:
        pregnancy:
          $ref: "#/components/schemas/Pregnancy"
        membership:
          $ref: "#/components/schemas/Membership"
    PregnancyDetail:
      type: object
      required: [pregnancy, members, myMembership]
      properties:
        pregnancy:
          $ref: "#/components/schemas/Pregnancy"
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberSummary"
        myMembership:
          $ref: "#/components/schemas/Membership"

    # ── Invitations ───────────────────────────────────────────────────────────
...