# lib/api-zod/src/generated/types/memberSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nest API specification
 * OpenAPI spec version: 0.2.0
 */

export interface MemberSummary {
  id: number;
  displayName: string;
  role: string;
}