# lib/api-zod/src/generated/types/pregnancyWithMembership.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nest API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Membership } from './membership';
import type { Pregnancy } from './pregnancy';

export interface PregnancyWithMembership {
  pregnancy: Pregnancy;
  membership: Membership;
}